+++ pkg/test/devstack/devstack_tests_old.go
@@ -1,4 +1,4 @@
-package test
+package devstack
 
 // import (
 // 	"context"
