+++ pkg/test/scheduler/scheduler_test.go
@@ -1,4 +1,4 @@
-package test
+package scheduler
 
 import (
 	"reflect"
